added more error handling for crud and endpoints, not yet fully tested

+++ sending-options-backend/src/helpers/crud.js
@@ -4,14 +4,23 @@
 */
 
 
+const mongoose = require('mongoose');
+const { ErrorHandler } = require('./error');
+
 // get all items from model
 async function getAll(request, response, next, model) {
     try {
         const resource = await model.find().exec();
+
+        // if no resource found, return error
         if (!resource) {
-            // return next(errors.RESOURCE_NOT_FOUND());
+            throw new ErrorHandler(400, 'Invalid request');
+        // if resource found, but empty, return status accordingly
+        } else if (resource.length === 0) {
+            return response.sendStatus(204);
         }
-        return response.json(resource);
+        // otherwise, return resource
+        return response.status(200).json(resource);
 
     } catch(error) {
         next(error);
@@ -21,11 +30,20 @@ async function getAll(request, response, next, model) {
 // get single item from model
 async function getSingle(request, response, next, model) {
     try {
-        const resource = await model.findById(request.params.id).exec();
-        if (!resource) {
-            //return next(errors.RESOURCE_NOT_FOUND())
+
+        // if id in params is not a valid object id, throw error
+        if (!mongoose.Types.ObjectId.isValid(request.params.id)) {
+            throw new ErrorHandler(404, 'Invalid ID');
         }
-        return response.json(resource);
+
+        const resource = await model.findById(request.params.id).exec();
+
+        // if no resource found, throw error
+        if (!resource) {
+            throw new ErrorHandler(404, 'No resouce for specified ID');
+        }
+        // otherwise, return resource
+        return response.status(200).json(resource);
 
     } catch(error) {
         return next(error);
@@ -33,13 +51,19 @@ async function getSingle(request, response, next, model) {
 }
 
 // create new item for model
-async function create(request, response, next, model) {
+function create(request, response, next, model) {
     try {
-        const resource = await model.create(request.body);
-        if (!resource) {
-            //return next(errors.RESOURCE_NOT_CREATED());
+        const resource = model.create(request.body);
+
+        // if no resource created, throw error
+        if (Object.keys(resource).length === 0) {
+            throw new ErrorHandler(400, 'Could not create item');
         }
-        return response.json({ message: resource._id +' created successfully' });
+        // otherwise, return status and newly created resource
+        return response.status(201).json({ 
+            message: 'Item created successfully', 
+            resource
+        });
 
     } catch(error) {
         return next(error);
@@ -49,12 +73,24 @@ async function create(request, response, next, model) {
 // update item from model
 async function update(request, response, next, model) {
     try {
-        const resource = await model.findByIdAndUpdate(request.params.id, request.body, { new: true }).exec();
-        if (!resource) {
-            // return next(errors.RESOURCE_NOT_FOUND());
+
+        // if id in params is not a valid object id, throw error
+        if (!mongoose.Types.ObjectId.isValid(request.params.id)) {
+            throw new ErrorHandler(404, 'Invalid ID');
         }
+
+        const resource = await model.findByIdAndUpdate(request.params.id, request.body, { new: true }).exec();
+
+        // if no resource found, throw error
+        if (!resource) {
+            throw new ErrorHandler(400, 'Could not find or update item');
+        }
+        // otherwise, save resource and return status and updated resource
         resource.save();
-        return response.json({ message: resource._id +' updated successfully' });
+        return response.status(201).json({ 
+            message: 'Item updated successfully', 
+            resource
+        });
 
     } catch(error) {
         return next(error);
@@ -64,11 +100,23 @@ async function update(request, response, next, model) {
 // remove item from model
 async function remove(request, response, next, model) {
     try {
-        const resource = await model.findByIdAndRemove(request.params.id).exec();
-        if (!resource) {
-            // return next(errors.RESOURCE_NOT_FOUND());
+
+        // if id in params is not a valid object id, throw error
+        if (!mongoose.Types.ObjectId.isValid(request.params.id)) {
+            throw new ErrorHandler(404, 'Invalid ID');
         }
-        return response.json({ message: resource._id +' deleted successfully' });
+
+        const resource = await model.findByIdAndRemove(request.params.id).exec();
+
+        // if no resource found, throw error
+        if (!resource) {
+            throw new ErrorHandler(400, 'Could not find or delete item');
+        }
+        // otherwise, return status and deleted resource
+        return response.status(200).json({ 
+            message: 'Item deleted successfully', 
+            resource
+        });
 
     } catch(error) {
         return next(error);
